Move addRights handling to activitypub.activity service
