+++ src/middleware/packages/activitypub/services/activitypub/subservices/activity.js
@@ -1,14 +1,17 @@
-const { ControlledContainerMixin } = require('@semapps/ldp');
-const { MIME_TYPES } = require('@semapps/mime-types');
 const { Errors: E } = require('moleculer-web');
+const { ControlledContainerMixin, hasType } = require('@semapps/ldp');
+const { MIME_TYPES } = require('@semapps/mime-types');
 const { objectCurrentToId, objectIdToCurrent, arrayOf } = require('../../../utils');
-const { PUBLIC_URI, FULL_ACTIVITY_TYPES } = require('../../../constants');
+const { PUBLIC_URI, FULL_ACTIVITY_TYPES, ACTIVITY_TYPES } = require('../../../constants');
+const ActivitiesHandlerMixin = require('../../../mixins/activities-handler');
 
 const ActivityService = {
   name: 'activitypub.activity',
-  mixins: [ControlledContainerMixin],
+  mixins: [ControlledContainerMixin, ActivitiesHandlerMixin],
   settings: {
     baseUri: null,
+    podProvider: false,
+    // ControlledContainerMixin settings
     path: '/as/activity',
     acceptedTypes: Object.values(FULL_ACTIVITY_TYPES),
     accept: MIME_TYPES.JSON,
@@ -103,6 +106,110 @@ const ActivityService = {
     after: {
       get(ctx, res) {
         return objectCurrentToId(res);
+      }
+    }
+  },
+  activities: {
+    addRights: {
+      match: '*',
+      priority: 1,
+      async onEmit(ctx, activity) {
+        const activityUri = activity['@id'] || activity.id;
+        const recipients = await ctx.call('activitypub.activity.getRecipients', { activity });
+
+        // When a new activity is created, ensure the emitter has read rights also
+        // Don't do that on podProvider config, because the Pod owner already has all rights
+        if (!this.settings.podProvider) {
+          if (!recipients.includes(activity.actor)) recipients.push(activity.actor);
+        }
+
+        // Give read rights to the activity's recipients
+        // TODO improve performances by passing all users at once
+        // https://github.com/assemblee-virtuelle/semapps/issues/908
+        for (const recipientUri of recipients) {
+          await ctx.call(
+            'webacl.resource.addRights',
+            {
+              resourceUri: activityUri,
+              additionalRights: {
+                user: {
+                  uri: recipientUri,
+                  read: true
+                }
+              },
+              webId: 'system'
+            },
+            {
+              meta: {
+                skipObjectsWatcher: true
+              }
+            }
+          );
+
+          // If this is a Create activity, also give rights to the created object
+          if (hasType(activity, ACTIVITY_TYPES.CREATE)) {
+            await ctx.call(
+              'webacl.resource.addRights',
+              {
+                resourceUri: typeof activity.object === 'string' ? activity.object : activity.object.id,
+                additionalRights: {
+                  user: {
+                    uri: recipientUri,
+                    read: true
+                  }
+                },
+                webId: 'system'
+              },
+              {
+                meta: {
+                  skipObjectsWatcher: true
+                }
+              }
+            );
+          }
+        }
+
+        // If activity is public, give anonymous read right
+        if (await ctx.call('activitypub.activity.isPublic', { activity })) {
+          await ctx.call(
+            'webacl.resource.addRights',
+            {
+              resourceUri: activityUri,
+              additionalRights: {
+                anon: {
+                  read: true
+                }
+              },
+              webId: 'system'
+            },
+            {
+              meta: {
+                skipObjectsWatcher: true // We don't want to trigger an Update
+              }
+            }
+          );
+
+          // If this is a Create activity, also give anonymous read right to the created object
+          if (hasType(activity, ACTIVITY_TYPES.CREATE)) {
+            await ctx.call(
+              'webacl.resource.addRights',
+              {
+                resourceUri: typeof activity.object === 'string' ? activity.object : activity.object.id,
+                additionalRights: {
+                  anon: {
+                    read: true
+                  }
+                },
+                webId: 'system'
+              },
+              {
+                meta: {
+                  skipObjectsWatcher: true // We don't want to trigger an Update
+                }
+              }
+            );
+          }
+        }
       }
     }
   }
