+++ src/middleware/packages/auth/services/jwt.js
@@ -13,7 +13,7 @@ module.exports = {
     const publicKeyPath = path.resolve(this.settings.jwtPath, 'jwtRS256.key.pub');
 
     if (!fs.existsSync(privateKeyPath) && !fs.existsSync(publicKeyPath)) {
-      console.log('JWT keypair not found, generating...');
+      this.logger.info('JWT keypair not found, generating...');
       if (!fs.existsSync(this.settings.jwtPath)) {
         fs.mkdirSync(this.settings.jwtPath);
       }
